move slack to clients module
