+++ busy_beaver/apps/retweeter/task.py
@@ -1,8 +1,7 @@
 from datetime import timedelta
 import logging
 
-from busy_beaver import chipy_slack
-from busy_beaver.clients import kv_store, twitter
+from busy_beaver.clients import chipy_slack, kv_store, twitter
 from busy_beaver.config import TWITTER_USERNAME
 from busy_beaver.extensions import db, rq
 from busy_beaver.models import ApiUser, PostTweetTask
