+++ busy_beaver/clients.py
@@ -1,13 +1,24 @@
-from .adapters import GitHubAdapter, KeyValueStoreAdapter, MeetupAdapter, TwitterAdapter
+from .apps.external_integrations.oauth_providers.slack import SlackOAuthFlow
+from .adapters import (
+    GitHubAdapter,
+    KeyValueStoreAdapter,
+    MeetupAdapter,
+    SlackAdapter,
+    TwitterAdapter,
+)
 from .config import (
     GITHUB_OAUTH_TOKEN,
     MEETUP_API_KEY,
+    SLACK_TOKEN,
     TWITTER_ACCESS_TOKEN,
     TWITTER_ACCESS_TOKEN_SECRET,
     TWITTER_CONSUMER_KEY,
     TWITTER_CONSUMER_SECRET,
+    SLACK_CLIENT_ID,
+    SLACK_CLIENT_SECRET,
 )
 
+chipy_slack = SlackAdapter(SLACK_TOKEN)  # Default Workspace -- this is being phased out
 github = GitHubAdapter(GITHUB_OAUTH_TOKEN)
 kv_store = KeyValueStoreAdapter()
 meetup = MeetupAdapter(MEETUP_API_KEY)
@@ -17,3 +28,5 @@ twitter = TwitterAdapter(
     TWITTER_ACCESS_TOKEN,
     TWITTER_ACCESS_TOKEN_SECRET,
 )
+
+slack_oauth = SlackOAuthFlow(SLACK_CLIENT_ID, SLACK_CLIENT_SECRET)
